+++ spec/lumberjack/sidekiq/attribute_passthrough_middleware_spec.rb
@@ -69,6 +69,32 @@ RSpec.describe Lumberjack::Sidekiq::AttributePassthroughMiddleware do
       end
     end
 
+    it "does not mutate a shared logging options hash" do
+      shared_logging = {"args" => ["arg1"], "attributes" => {"env" => "test"}}.freeze
+      shared_logging["attributes"].freeze
+      job["logging"] = shared_logging
+
+      logger.tag(user_id: 123) do
+        middleware.call("MyWorker", job, "default", nil) do
+          expect(job.dig("logging", "attributes")).to eq("env" => "test", "user_id" => 123)
+          expect(job["logging"]).not_to equal(shared_logging)
+        end
+      end
+
+      other_job = {"args" => [], "logging" => shared_logging}
+      logger.tag(request_id: "abc") do
+        middleware.call("MyWorker", other_job, "default", nil) do
+          expect(other_job.dig("logging", "attributes")).to eq("env" => "test", "request_id" => "abc")
+        end
+      end
+    end
+
+    it "does not add logging options to the job if there are no attributes to pass through" do
+      middleware.call("MyWorker", job, "default", nil) do
+        expect(job).to eq("args" => [1, 2, 3])
+      end
+    end
+
     it "does not add attributes if the logger is not a Lumberjack logger" do
       allow(Sidekiq).to receive(:logger).and_return(Logger.new(StringIO.new))
       logger.tag(user_id: 123, request_id: "abc") do